+++ src/Knp/Bundle/KnpBundlesBundle/DataFixtures/ORM/Data.php
@@ -153,6 +153,11 @@ EOT;
                 'tags'          => ($i%2) ? array('1.0', '1.1') : array(),
                 'usesTravisCi'  => ($i%2) ? false : true,
                 'composerName'  => ($i%2) ? null : 'knplabs/knp-menu-bundle',
+                'symfonyVersions' => array(
+                    'dev-master' => '2.1.*',
+                    '1.2.0' => '2.0.*',
+                    '1.1.0' => '2.*',
+                ),
                 'state'         => $states[mt_rand(0, 3)],
                 'travisCiBuildStatus'  => ($i%2 == 0) ? $trilean[$i%3] : null,
                 'nbFollowers'   => $i*10,