+++ CHANGELOG.markdown
@@ -1,2 +1,4 @@
+* 2012-08-20
+  * BC: Bundle entity property symfonyVersion string turned into symfonyVersions array
 * 2012-08-15
   * BundleActivityTwigExtension renamed to BundleUtilsExtension